test: migrate cron-process-usage-events to bdd
